ButtsBlazor.Generator2/requirements.txt
-i https://pypi.org/simple
accelerate==0.23.0
datasets==2.14.5
diffusers==0.21.4
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.1
tensorrt
triton==2.0.0; sys_platform != "win32"
tqdm==4.66.1
transformers==4.33.2
xformers==0.0.22
gradio